type of resources and radius changed to qreal

+++ defaultplayerintelligence.cpp
@@ -24,7 +24,7 @@ void DefaultPlayerIntelligence::think(const GameTime &gameTime)
         if (enemyPlayer->planets().isEmpty()) return; // enemy has no planets!
         Planet *enemyPlanet = *Random::randomElement(enemyPlayer->planets());
         // attack enemy planet with 25-75% of my choosen planets resources
-        int res = int(myPlanet->resources() * Random::randomReal(0.25, 0.75));
+        int res = myPlanet->resources() * Random::randomReal(0.25, 0.75);
         myPlanet->setResources(myPlanet->resources() - res);
         Ship *ship = new Ship(myPlanet->position(), enemyPlanet, res, myPlanet->color(), self);
         self->ships().insert(ship);
